requirements.txt
Pillow==5.2.0
tqdm==4.25.0
